+++ juntagrico/view_decorators.py
@@ -1,5 +1,6 @@
 from functools import wraps
 
+from django.contrib.auth.decorators import user_passes_test
 from django.shortcuts import get_object_or_404, redirect
 
 from juntagrico.models import Subscription
@@ -37,3 +38,15 @@ def create_subscription_session(view):
         return response
 
     return wrapper
+
+
+def any_permission_required(*perms):
+    """
+    Decorator for views that checks whether a user has any of the given permissions
+    """
+    def check_perms(user):
+        # check if the user has any of the permission
+        if set(user.get_all_permissions()) & set(perms):
+            return True
+        return False
+    return user_passes_test(check_perms)
